requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
drf-nested-routers==0.93.3
idna==3.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
urllib3==1.26.7
